third_party/blink/web_tests/http/tests/devtools/components/split-string-by-regexes-expected.txt

Tests TextUtils.TextUtils.splitStringByRegexes.


Running: testSimple
"hello", 0, 0
"123", 5, 1
"hello", 8, 0
"123", 13, 1

Running: testMatchAtStart
"yes", 0, 0
" thank you", 3, -1

Running: testMatchAtEnd
"yes thank ", 0, -1
"you", 10, 0

Running: testAvoidInnerMatch
"image: ", 0, -1
"url("red.com")", 7, 0

Running: testNoMatch
"nothing", 0, -1

Running: testNoMatches
"nothing", 0, -1

Running: testComplex
"Start. ", 0, -1
"(okay)", 7, 0
" kit-", 13, -1
"ka", 18, 2
"t ", 20, -1
"okay", 22, 1
" ", 26, -1
"(kale)", 27, 0
" ", 33, -1
"ka", 34, 2
"( )", 36, 0
" ", 39, -1
"okay", 40, 1
". End", 44, -1
